Feature/deploy (#27)
* Create timelock for public sale

* Add initializer with mint

* fix merge issue

* Add a test verifying the upgrade pattern

* add deployment stuff

* comment out networks where no key is checked into source
