+++ data/opensea_metadata.json
@@ -1,8 +1,8 @@
 {
-    "name": "Rate Test Contract",
+    "name": "RarePizzas Reservations",
     "description": "Pizza ipsum dolor amet meatball garlic sauce green bell peppers white pizza lasagna pepperoni, spinach anchovies. Bacon string cheese pesto, red onions peppers steak marinara party sauteed onions pepperoni garlic philly chicken bacon & tomato.",
     "image": "https://uploads-ssl.webflow.com/603ff4e71a58e224f91525c8/603ff5b51acddaba7406268d_rarepizzas_paaaaaarttttttttyyyyyyyyy.gif",
     "external_link": "rarepizzas.com",
     "seller_fee_basis_points": 500,
-    "fee_recipient": "0xPIZZA"
+    "fee_recipient": "0x102d3B5ca9C8675C20B2F7E2171B3ecDcbe3Fc82"
 }
