+++ hardhat.config.ts
@@ -13,6 +13,19 @@ task('accounts', 'Prints the list of accounts', async (args, hre) => {
   }
 })
 
+// Go to https://www.alchemyapi.io, sign up, create
+// a new App in its dashboard, and replace "KEY" with its key
+const ALCHEMY_API_KEY = "KEY";
+
+// Replace these private keys with your testnet account private keys
+// To export your private key from Metamask, open Metamask and
+// go to Account Details > Export Private Key
+// Be aware of NEVER putting real Ether into testing accounts
+const GOERLI_PRIVATE_KEY = "YOUR GOERLI PRIVATE KEY";
+const RINKEBY_PRIVATE_KEY = "YOUR RINKEBY PRIVATE KEY";
+const ROPSTEN_PRIVATE_KEY = "YOUR ROPSTEN PRIVATE KEY";
+
+
 // You need to export an object to set up your config
 // Go to https://hardhat.org/config/ to learn more
 
@@ -26,7 +39,23 @@ task('accounts', 'Prints the list of accounts', async (args, hre) => {
 const config: HardhatUserConfig = {
   defaultNetwork: 'hardhat',
   networks: {
+    // goerli: {
+    //   url: `https://eth-goerli.alchemyapi.io/v2/${ALCHEMY_API_KEY}`,
+    //   accounts: [`0x${GOERLI_PRIVATE_KEY}`]
+    // },
     hardhat: {},
+    // mainnet: {
+    //   url: `https://eth-mainnet.alchemyapi.io/v2/${ALCHEMY_API_KEY}`,
+    //   accounts: [`0xNOPE!`]
+    // },
+    // rinkeby: {
+    //   url: `https://eth-rinkeby.alchemyapi.io/v2/${ALCHEMY_API_KEY}`,
+    //   accounts: [`0x${RINKEBY_PRIVATE_KEY}`]
+    // },
+    // ropsten: {
+    //   url: `https://eth-ropsten.alchemyapi.io/v2/${ALCHEMY_API_KEY}`,
+    //   accounts: [`0x${ROPSTEN_PRIVATE_KEY}`]
+    // }
   },
   solidity: {
     version: '0.8.0',
